+++ package.json
@@ -12,6 +12,7 @@
     "@formkit/auto-animate": "^0.8.2",
     "@tailwindcss/vite": "^4.1.11",
     "axios": "^1.11.0",
+    "lodash.debounce": "^4.0.8",
     "tailwindcss": "^4.1.11",
     "vue": "^3.5.12",
     "vue-router": "^4.5.1"
