+++ src/components/Drawer.vue
@@ -1,19 +1,48 @@
 <script setup>
-  import { inject } from "vue";
-
-  const { closeDrawer } = inject("cart");
+  //const { closeDrawer } = inject("cart");
 
   import DrawerHead from "./DrawerHead.vue";
   import CartItemList from "./CartItemList.vue";
   import infoBlock from "./infoBlock.vue";
+  import { inject, ref, computed } from "vue";
+  import axios from "axios";
 
-  defineProps({
+  const props = defineProps({
     totalPrice: Number,
     vatPrice: Number,
-    buttonDisabled: Boolean,
   });
 
-  const emit = defineEmits(["createOrder"]);
+  const isCreating = ref(false);
+  const orderId = ref(null);
+
+  const { cart, closeDrawer } = inject("cart");
+
+  const buttonDisabled = computed(
+    () => isCreating.value || cart.value.length === 0
+  );
+
+  const createOrder = async () => {
+    try {
+      isCreating.value = true;
+      const { data } = await axios.post(
+        `https://c54d42806c01eb8f.mokky.dev/orders`,
+        {
+          items: cart.value,
+          totalPrice: props.totalPrice.value,
+        }
+      );
+
+      orderId.value = data.id;
+
+      cart.value = [];
+
+      return data;
+    } catch (err) {
+      console.log(err);
+    } finally {
+      isCreating.value = false;
+    }
+  };
 </script>
 
 <template>
@@ -28,10 +57,17 @@
     >
       <DrawerHead />
 
-      <CartItemList v-if="totalPrice" />
+      <CartItemList v-if="totalPrice && !orderId" />
 
       <infoBlock
-        v-else
+        v-if="!totalPrice && orderId"
+        title="Заказ оформлен!"
+        :description="`Ваш заказ #${orderId} скоро будет передан курьерской доставке`"
+        image-url="/order-success-icon.png"
+      />
+
+      <infoBlock
+        v-if="!totalPrice && !orderId"
         title="Корзина пустая"
         description="Добавьте хотя бы одну пару кроссовок, чтобы сделать заказ."
         image-url="/package-icon.png"
@@ -54,7 +90,7 @@
 
         <button
           :disabled="buttonDisabled"
-          @click="() => emit('createOrder')"
+          @click="createOrder"
           class="bg-lime-500 p-3 w-full text-white rounded-xl duration-300 hover:bg-lime-600 active:bg-lime-700 disabled:bg-slate-300 cursor-pointer mt-5"
         >
           Оформить заказ
